Correcciones en el controlador de Citas y ClaseDAO
Se modificaron el controlador y la clase DAO de Cita para ser compatibles con el ultimo commit de esta rma

+++  ClinicaAPI/Controllers/CitaController.cs
@@ -37,13 +37,13 @@ public class CitaController : Controller
         return Ok(mensaje);
     }
     [HttpDelete("eliminarCita/{id}")]
-    public async Task<ActionResult> eliminarCita(int id)
+    public async Task<ActionResult> eliminarCita(long id)
     {
         await Task.Run(() => new CitaDAO().eliminarCita(id));
         return Ok();
     }
     [HttpGet("buscarCita/{id}")]
-    public async Task<ActionResult> buscarCita(int id)
+    public async Task<ActionResult> buscarCita(long id)
     {
         var lista = await Task.Run(() => new CitaDAO().buscarCita(id));
         return Ok(lista);
